Fix axios mock and move commit type
