--- src/types/axios.d.ts
@@ -1,1 +1,0 @@
-declare module 'axios/lib/adapters/http';
